Вывод интерфейс в отдельный файл.
