+++ src/SuperBug.Politrange.Services/Persons/PersonService.cs
@@ -4,15 +4,6 @@ using SuperBug.Politrange.Models;
 
 namespace SuperBug.Politrange.Services.Persons
 {
-    public interface IPersonService
-    {
-        IEnumerable<Person> GetAll();
-        Person GetById(int id);
-        Person Add(Person person);
-        bool Update(Person person);
-        bool Remove(int id);
-    }
-
     public class PersonService: IPersonService
     {
         private readonly IKeywordRepository keywordRepository;
